dom/media/mediasource/test/test_BufferedSeek_mp4.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: seeking in buffered range</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(function(ms, v) {
  ms.addEventListener("sourceopen", function() {
    const sb = ms.addSourceBuffer("video/mp4");

    fetchWithXHR("bipbop/bipbop2s.mp4", function(arrayBuffer) {
      sb.appendBuffer(new Uint8Array(arrayBuffer));
    });

    const target = 1.3;

    v.addEventListener("loadedmetadata", function() {
      ok(true, "received loadedmetadata");
      v.currentTime = target;
    });

    let wasSeeking = false;

    v.addEventListener("seeking", function() {
      wasSeeking = true;
      is(v.currentTime, target, "Video currentTime at target");
    });

    v.addEventListener("seeked", function() {
      ok(wasSeeking, "Received expected seeking and seeked events");
      is(v.currentTime, target, "Video currentTime at target");
      SimpleTest.finish();
    });
  });
});

</script>
</pre>
</body>
</html>
